refact: useSearch hook을 별도의 파일로 추출
